Start Piece practicing, better layout, form bugs
I started to add practicing to the piece pages and save the practice
sessions in the database (models, queries, pages, etc.) Added Random and
Sequenced Practicing of piece spots as well as random starting points

I changed the handling of the page nav and title to be more consisted
with two different elements depending on page size.

I fixed some bugs in the form that were preventing values from being
saved. valueAsNumber in the register function options was crucial.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>