Remember, titles, new stage, no spots bug
Fixed a bug where you could not user the add spots page for a piece that
has no spots.

Added title to hxrender so that pages have the same title when swapped
by htmx and when loaded fresh.

Added a remember me checkbox to the login page so users can be quickly
logged back in with their passkey.

Added a more repeat practicing stage for spots that need extra repeat
practicing and a dialog to choose which stage to promote to.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>