Spot display on repeat practice, component registration errors
There is now a spot practice display component that can be reused and is
displayed on the repeat practice page.

Each component registration is now wrapped in its own try catch block so
they don't stop each other from registering. The reminders display is
now in main because it is needed in multiple places.

closes #5

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>