Random Practicing Evaluation, bug fixes
Fixed a bug where practicing a spot more than once in a practice session
would cause the session to not save (Unique constraint failed). This added a 'reps' column to the
practice_session_spot table.

Random practicing self-evaluation was added with 'excellent', 'poor',
and 'fine'. The system will now make promotion/demotion recommendations
based on how well it feels the spots are going. This required adding a
stage started column to track when the spot started in its stage so the
system can tell if a spot has been practice enough or is failing to
improve.

Closes #31
Closes #57

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>