Add persistence, resumability, and delete to practice plans
The active practice plan is now tracked in the database on the user
(which we are querying anyway) along with a start time for expiration
purposes.

Practice plans can now be resumed if they are less than five hours old.

The actice practice plan can now be stopped and practice plans can also
be deleted. The active plan being deleted is handled, even though the
button isn't there.

closes #48

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>